Add box_future and box_future_middleware functions to helpers
Introduce two new utility functions, box_future and box_future_middleware, to the helpers module for easier handling of futures in middleware. Update relevant imports in types and app modules to utilize these new functions, enhancing code organization and readability.
